Move functionality to steps file
